fixes #6981 - handle non-yum sync statuses
Splitting out the sync presenter into a few different files
depending on the type
